move colorByExecution handler to utils
